Fix 32 bit build again

+++ Qbert/LevelComponent.cpp
@@ -21,7 +21,7 @@ bool dae::LevelComponent::NextRound() {
 
 	// If it's larger or equal to the rounds per level,
 	// reset it, go to the next level, and notify the subject
-	if (m_round >= m_levelInfo.rounds.size()) {
+	if (m_round >= static_cast<int>(m_levelInfo.rounds.size())) {
 		for (auto& command : m_nextLevelCommands) {
 			command->Execute();
 		}
